refactor: ResponseEntity, 엔드포인트 수정, 회원 / 관리자 구조 분리
