+++ backend/src/main/java/com/team5/backend/domain/member/admin/entity/Admin.java
@@ -1,0 +1,4 @@
+package com.team5.backend.domain.member.admin.entity;
+
+public class Admin {
+}
