+++ backend/src/main/java/com/team5/backend/domain/member/member/entity/Role.java
@@ -1,4 +1,4 @@
-package com.team5.backend.domain.member.entity;
+package com.team5.backend.domain.member.member.entity;
 
 import lombok.Getter;
