+++ backend/src/main/java/com/team5/backend/domain/review/entity/Review.java
@@ -1,6 +1,6 @@
 package com.team5.backend.domain.review.entity;
 
-import com.team5.backend.domain.member.entity.Member;
+import com.team5.backend.domain.member.member.entity.Member;
 import com.team5.backend.domain.product.entity.Product;
 import jakarta.persistence.*;
 import lombok.*;
